+++ RELEASE_NOTES.md
@@ -41,7 +41,7 @@
 - **Uploads**: Fixed issue with uploading files with filename length smaller than three characters (e.g. "24" - Damn you, Jack!).
 - **Table Asset Manager / Playlists**: Fixed asset search for playlists. We somehow forgot that. You can now search the frontend's asset database for media for your playlists, e.g. "music".
 - **Highscore Parsing**: Added additional lowercase check for VPReg.stg based highscores ("HELLBOY" problem).
-- **Highscore Parsing**: Added test coverage for over 50 new text file based highscores (thanks to @gorgatron) and added the missing parsing support for some of them.
+- **Highscore Parsing**: Added test coverage for over 50 new text file and VPReg.stg file based highscores (thanks to @gorgatron) and added the missing parsing support for some of them.
 - **System Manager / Visual Pinball**: Fixed issue that the parent folder of the "Tables" folder was used for the VPX system manager component (instead of the actual installation folder). Some users have selected a different folder just for the tables.
 - **Table Data Manager**: Fixed dialog sizing issues.
 - **Backglass Data Exporter**: